requirements.txt
alembic==1.0.5
aniso8601==4.0.1
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
Authlib==0.10
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
coverage==4.5.2
cryptography==2.4.2
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Migrate==2.3.1
flask-restplus==0.12.1
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.1.0
marshmallow==2.16.3
more-itertools==4.3.0
numpy==1.15.4
pluggy==0.8.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
py==1.7.0
pycparser==2.19
pytest==4.0.1
pytest-cov==2.6.0
python-dateutil==2.7.5
python-dotenv==0.10.0
python-editor==1.0.3
pytz==2018.7
requests==2.21.0
six==1.12.0
SQLAlchemy==1.2.14
urllib3==1.24.1
Werkzeug==0.14.1
